chore(error codes): moves the error code constants PE-624
