+++ src/CLICommand/cli_command.ts
@@ -3,7 +3,7 @@ import { program } from 'commander';
 import { CLIAction } from './action';
 import { CliApiObject, ParsedParameters } from './cli';
 import { Parameter, ParameterName, ParameterOverridenConfig } from './parameter';
-import { ERROR_EXIT_CODE } from './constants';
+import { ERROR_EXIT_CODE } from './error_codes';
 
 export type CommandName = string;
 export interface CommandDescriptor {
