--- src/CLICommand/constants.ts
@@ -1,2 +1,0 @@
-export const SUCCESS_EXIT_CODE = 0;
-export const ERROR_EXIT_CODE = 1;
