+++ src/CLICommand/parameters_helper.test.ts
@@ -29,7 +29,7 @@ import '../parameter_declarations';
 import { urlEncodeHashKey } from '../utils';
 import { stubArweaveAddress } from '../utils/stubs';
 import { CLIAction } from './action';
-import { SUCCESS_EXIT_CODE } from './constants';
+import { SUCCESS_EXIT_CODE } from './error_codes';
 
 const expectedArweaveAddress = stubArweaveAddress('P8aFJizMVBl7HeoRAz2i1dNYkG_KoN7oB9tZpIw6lo4');
