+++ src/commands/create_folder.ts
@@ -9,7 +9,7 @@ import {
 import { arDriveFactory } from '..';
 import { Wallet } from '../wallet_new';
 import { FeeMultiple } from '../types';
-import { SUCCESS_EXIT_CODE } from '../CLICommand/constants';
+import { SUCCESS_EXIT_CODE } from '../CLICommand/error_codes';
 import { CLIAction } from '../CLICommand/action';
 
 new CLICommand({
