+++ src/commands/generate_seedphrase.ts
@@ -1,7 +1,7 @@
 import { cliWalletDao } from '..';
 import { CLICommand } from '../CLICommand';
 import { CLIAction } from '../CLICommand/action';
-import { SUCCESS_EXIT_CODE } from '../CLICommand/constants';
+import { SUCCESS_EXIT_CODE } from '../CLICommand/error_codes';
 
 new CLICommand({
 	name: 'generate-seedphrase',
