+++ src/commands/get_address.ts
@@ -1,6 +1,6 @@
 import { CLICommand, ParametersHelper } from '../CLICommand';
 import { CLIAction } from '../CLICommand/action';
-import { SUCCESS_EXIT_CODE } from '../CLICommand/constants';
+import { SUCCESS_EXIT_CODE } from '../CLICommand/error_codes';
 import { SeedPhraseParameter, WalletFileParameter } from '../parameter_declarations';
 
 new CLICommand({
