+++ src/commands/get_mempool.ts
@@ -1,6 +1,6 @@
 import { CLICommand } from '../CLICommand';
 import { CLIAction } from '../CLICommand/action';
-import { SUCCESS_EXIT_CODE } from '../CLICommand/constants';
+import { SUCCESS_EXIT_CODE } from '../CLICommand/error_codes';
 import { fetchMempool } from '../utils';
 
 new CLICommand({
