+++ src/commands/list_all_drives.ts
@@ -2,7 +2,7 @@ import { ArFSDriveEntity } from 'ardrive-core-js';
 import { arDriveAnonymousFactory, cliWalletDao } from '..';
 import { CLICommand, ParametersHelper } from '../CLICommand';
 import { CLIAction } from '../CLICommand/action';
-import { SUCCESS_EXIT_CODE } from '../CLICommand/constants';
+import { SUCCESS_EXIT_CODE } from '../CLICommand/error_codes';
 import { AddressParameter, DrivePrivacyParameters } from '../parameter_declarations';
 
 new CLICommand({
